[IMP] account_reports: improved display of draft entries
Before that, when you toggled the option to see the unposted entries in accounting reports, they'd appear but, unlike in the rest of Odoo, they wouldn't stand out.
They now have the same styling as in list view

Also, a warning would appear in the aged receivable/payable report if some draft entries would be existing in the selected period although the option to see those wasn't present.

task-4443453

closes odoo/enterprise#76383

Related: odoo/odoo#192700
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Co-authored-by: hupo-odoo <hupo@odoo.com>