[FIX] l10n_nz_eft: test date
Fix a small issue with the test where the batch is created before the
freeze time, causing the batch date & name to be wrong.

The freeze time is instead removed, and the few
places which were dependant of the current date
(where the batch creation date matters) are
changed to dynamically expect today's date.

closes odoo/enterprise#76479

X-original-commit: 3d827b6c2d7f5c83785ecb6821092f349fca922f
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>