[FIX] l10n_it_pos: force load the current company's partner
Before this commit, it could happen that the partner associated
to the current company was not automatically loaded at the start
of the POS session.
Since this parter is required for proper operation of l10n_it_pos,
it could lead to unexpected errors.

After this commit, the partner is explicitly force-loaded at the
start of the POS session to ensure proper operation.

opw-4394672

X-original-commit: 2eb7469bc96497f446f9eeaa3fb8d4488df6e09b
Part-of: odoo/enterprise#77006
Related: odoo/odoo#193352
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>