[FIX] account_bank_statement_import: fix drag and drop
Before this commit, if you drag and drop a partner name from the form view of bank_reconciliation in the file import dropzone, a traceback is raise.

This commit fix this issue by hiding the dropzone when user is not dragging a file.

opw-4366605

closes odoo/enterprise#78037

closes odoo/enterprise#78610

X-original-commit: b1ef2b843aefcee246cc74c93fd3e9f58493bb0e
Related: odoo/odoo#195506
Related: odoo/odoo#196622
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>