[IMP] *: remove units of measure categories
Previous to this commit, units were classified into categories such as
length, weight, etc. If the user defined a base unit for a product, any
movement on that product (stock move, order line, etc.) can be only done
in a unit in the same category of the base unit of the product. This
commit removes unit categories and the previously mentioned enforcement.
This is useful for some business cases that need more flexibility (for
example: buying from vendors in pieces and selling in kilograms, etc.).
To allow for some structure, a new field is added to uom.uom model to
define a reference unit for any unit. Therefore, users can define a
hierarchy of units. A side effect of this change is that the `uom_uom`
model is restructured completely to be much simpler. The fields
`uom_type`, `color`, `factor_inv`, `ratio` are no longer needed. Each
`uom_uom` record now has only two fields that define its quantity:
`relative_factor` which is the quantity that this unit contains of its
reference unit (if exists), and `factor` which is the absolute quantity
of this unit all the way up to the most ancestor unit. The latter cannot
be controlled by users and not shown in any view but used in all
quantity computations. Another side effect related to point of sale app
is that the field `is_pos_groupable` is now set on the unit level
instead of category level. The value of this field is set in
odoo/upgrade#6629 to preserve the current behavior for users. Products
also now have a new field `uom_ids` labelled as "Packagings" where users
can define the units that this product can be sold/moved in.

Task-4252043

X-original-commit: d23ca81b24d8ddba873b2deba32b82f631a73390
Part-of: odoo/enterprise#77092
Related: odoo/odoo#193490
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>