[FIX] web_gantt: aggregation with subdivided columns
When the scale precision is half or quarter, the number of grid columns is
this.columnCount * this.model.metaData.scale.cellPart + 1 and not
this.columnCount + 1 as was assumed in aggregatePills. This led to a bad
computation of aggregates in columns sufficiently far on the right of the gantt
view in precision half or quarter.
Note that this error was not detected before because many gantt views do not use
that precisions or split groups anyway (via shouldMergeGroups).

closes odoo/enterprise#76654

X-original-commit: ffe7290a5a295a74d0ea13063346ed73136b44a6
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>