[IMP] *: unfiy rounding value on all units of measure
Previous to this commit, users could define a different rounding
precision on each unit of measure. This led to some incosistencies with
the decimal accuracy record "Product Unit of Measure" which is also used
in some scenarios to set the rounding precision of unit-related
computations. This commit removes the ability to set the rounding
precision on the unit level and makes it a computed field always based
on the above mentioned `decimal.precision` record, so that both values
are always the same. The field is now considered dummy and can be
removed in a later changed and all of its occurrences can be replaced
with the corresponding decimal accuracy record.

Task-4252043

X-original-commit: e9f30fbba12133bbcd1f54469d4b778989a56c53
Part-of: odoo/enterprise#77092
Related: odoo/odoo#193490
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>