[IMP] *: rename all labels of units of measure to "Unit"
This commit unifies all the appearances of the fields related to
`uom_uom` model to be "Unit" instead of "UoM", "Unit of Measure", etc.
This commit also adds new widget to be used in many2x fields that are
related to `uom_uom` model. The widget shows the relative factor and
reference unit in muted text next to the unit name in the autocomplete
dropdown.

Task-4252043

X-original-commit: 4417f4ea5a49d82670f70efa4ee7234b9a1d3930
Part-of: odoo/enterprise#77092
Related: odoo/odoo#193490
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>