[IMP] *: clean units of measure data
This commit archives some of the `uom_uom` records that are
automatically defined on `uom` module installation. This is mainly to
improve the `uom_uom` list view and make it more readable. Also some of
the archived imperial units are automatically activated on installing
the US localization. The commit also adds a new half-dozen unit labelled
"Pack of 6".

Task-4252043

X-original-commit: 284befa08755039c36559114149e8eb47e7ef897
Part-of: odoo/enterprise#77092
Related: odoo/odoo#193490
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>