[FIX] account_reports: sections tour
The sections' tour was breaking with the change of year.

We use the generic tax report for doing the testing. The generic tax
report has a periodicity of 1 month and opens at the previous period.

Our tests where based on the current date year. It works fine except on
the change of year. Indeed, we will be on January of year XXXX but the
report will open on December of year YYYY.

Now, we correctly remove 1 month to our dates to make sure we have the
correct year.

We also took the time to replace the usage of Date of plain javascript
with DateTime of the luxon library.

closes odoo/enterprise#76902

X-original-commit: 9b3c8f6357d5b832299d2ff2c5fb4f92e3c8286e
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>