[IMP] sale_subscription: avoid locking the entire transaction row
Before this commit, the transaction row was locked when payment was
performed in the subscription invoice cron.
Using `SELECT FOR UPDATE` locks the entire row, primary key included (id
in our case). As a result, it could prevent insertion or update or other
table refering to the id too. As it is not necessary and could trigger
unnecessary deadlock, it is better to rely on `SELECT FOR NO KEY UPDATE`
which allow the update of foreign keys.

https://www.postgresql.org/docs/16/explicit-locking.html#LOCKING-ROWS

See https://github.com/odoo/enterprise/pull/74078#discussion_r1873729342

closes odoo/enterprise#76866

Taskid: 4391755
X-original-commit: 07f01f39cb46f1763637b50e628dfb613b0d2266
Signed-off-by: Olivier Dony (odo) <odo@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>