[IMP] website_appointment: support turnstile
The previous commit implements captcha for this form

Turnstile can be implemented on top of it by just invoking the turnstile script in js.

As it's implicitly checked when recaptcha is checked, this also avoids any issue with the form potentially not working if turnstile is enabled.

task-4335141

closes odoo/enterprise#81586

X-original-commit: c08cd03bb7f8c05536df9930be75dba4f27f77cd
Related: odoo/odoo#202078
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>