[FIX] hr_contract_salary: keep benefits values order
Since the Select owl componenent is used to display the benefits values,
the 'options' of the select are displayed alphabetically. It is caused
by the default value to 'true' for the autoSort attribute of the
component.

As we give the data in a specific order to the select in salary
cofigurator we don't want the JS to re-order it.

e.g.: the cars are sorted by cost and we don't wan't it in the
alphabetical order.

closes odoo/enterprise#76747

X-original-commit: 70768e272e81aa2929713217e93172d863e72e17
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>