[FIX] planning: wrong shift split after copying previous week
Steps to reproduce:
-------------------
1. Install the Planning App
2. Pick a week and assign a shift from Monday to Friday for a resource A
3. Create two time offs for resource A: one on Tuesday and one on Thursday of the week after you created the shift
4. Now, in the planning gantt view, move to the next week (the week after you created the shift)
5. Use the copy previous week action
6. The copied shifts are split in a way they shouldn't. On Monday, Wednesday and Friday, there are two copied slots instead of just one.

Fix:
-------------------
Currently, the _merge_slots_values() method is called when we use copy previous week on a forecasted slot (> 24 hours), which is the case here.
This method takes work intervals and unavailabilites in parameters, and returns merged slots by taking into account the unavailabilites.
To fix the issue mentioned above, we removed a condition that only allows slots to be merged if their combination creates a forecasted slot (> 24 hours).
But we argue that we should also be able to merge smaller slots (< 24 hours).

task-4368919
version-18.0

closes odoo/enterprise#78637

X-original-commit: c98e9186023eb0b885322925337b2047086465fa
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Maxime de Neuville (mane) <mane@odoo.com>