[REF] *: company service is no longer a service
This commit aims to simplify the way developers can access company
related information in the web client. Before this commit, in
components, they had to import the `userService` hook and do something
like `this.company = userService("company")` in the setup. In services,
they had to explicitly declare the dependency on the company service. In
other functions (like helpers), it was a bit painful: the function had
to take an env in arguments, which wasn't ideal.

This commit makes it easier to access the company information. The
company service has been removed. Instead, we use the existing `user`
object which can be imported directly from anywhere. The `user` object
was chosen because the company information is all related to the
currently logged-in user.

For clarity, the company information has been renamed:

user.allowedCompanies: The list of companies that the user is allowed to
    log in to. Note that any of the other companies in the following
    keys will always be in this list.
user.defaultCompany: The default company to log in to, if the cookies
    are empty.
user.activeCompanies: The list of the companies that the user is
    currently logged in to
user.activeCompany: The main company that the user is currently logged
    in to. This is the company displayed in the main navigation bar in
    the web client.

Note that, something similar has already been done for the `user`
service, the `cookie` service, the `router` service, and the `rpc`
service.

This commit also allows the `user` object to take ownership of the
company related information in the session. The problem before this
commit was that there were 2 sources of truth for company related
information : the company service and the `session_info`. The
information is sent from the server to the client within the page (in
`session_info`). Then the company service was built using this
information. This means that the developer could access the information
in 2 different ways. Now, the `user` object takes the full ownership of
these company related keys by removing them from the session after
reading them. Note that, something similar is already done for the user
related information and for the currencies (see currency.js).

Part-of task-id 3621046
Part-of task-id 4250356
Part-of task-id 2224776

closes odoo/enterprise#76685

Related: odoo/odoo#192847
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>