[FIX] {helpdesk_}timesheet{_grid}: display timer running on grid row when start timer
Before this commit, when the user starts a new timer on the grid view
and a timesheet is created, the timer is not marked as running on the
row contained the project. Same issue when a timer is running and the
user refreshes the page.

This commit adds the needed data to make sure the grid view finds which
row has the timer running.

closes odoo/enterprise#76982

X-original-commit: f3940ce65be668c5133558ffa60f33eca52edb3b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>