[FIX] web_studio: prevent crash when studio SysTray re-renders
Following the "nice urls" [task][1]
It may happen that the studio's systray item is re-rendered while the actionService is loading
a URL with multiple actions

In this case, there may be a crash because the SystrayItem would check if the current action
(in this case from the virtual controller) is editable.

After this commit, there is no crash.

part of task-4391729

[1]: odoo/odoo@c63d14a0485a553b74a8457aee158384e9ae6d3f

closes odoo/enterprise#76757

X-original-commit: 2e8db5d86afbd150634be2a1a467e0338bf596d3
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>