[IMP] *: add t-autoprefix for assets bundle
*: account_followup,account_reports,mrp_account_enterprise,
   sign,snailmail_account_followup,web_studio

This commit allows to only enable when needed the "autoprefixing"
feature of CSS rules used for compatibility with (very) old browsers. To
do so, it introduces the boolean `t-autoprefix` attribute on `<t
t-call-assets="..."/>` in QWeb's templates.

As those prefixes (e.g. `-webkit-`, `-moz-`...) are basically only
required for wkhtmltopdf compatibility nowadays, only the reports'
bundles stil use it.

Also disabling it for the other bundles (backend, frontend...) result
into a slight gain in the bundles' size (+/- 20ko uncompressed on a
database "all").

closes odoo/enterprise#76812

Related: odoo/odoo#193043
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>