[FIX] stock_barcode: avoid empty tracking number list
Main:
Encoding a SGTIN EPC requires a tracking number to make the tag unique.
When the product is not tracked, a "fake" tracking number is used.
Otherwise, we use the tracking number registered on the Move Line.
However, depending on the state of the Move Line, lot_id may not be set
and the tracking number will rather be available in lot_name.
Currently, we only rely on lot_id, which may lead to an empty tracking
number list, which in turn raises an unhandled exception.

 - We should try getting the tracking number from lot_name when lot_id
 is not set;
 - We should handle the case where the tracking number list is empty.

N.B. : Only an empty tracking number list is problematic, the case of a
list populated with null value is already handled by the epc_encoder.

Before:
 - We only get the tracking number from lot_id.name;
 - No verification is done on the tracking number list.

After:
 - Default to lot_name when lot_id is not set.
 - When the tracking number list is empty, set an error warning as the
 electronic_product_code on the concerned move lines.

To reproduce:
 - Create a Product (A) tracked by serial with a barcode
 - Create a receipt Picking for 30 A
 - "Mark as Todo" the picking
 - Click on the smart button "Moves"
 - Select all the lines
 - Export the data with 'Electronic Product Code' field selected

closes odoo/enterprise#78662

X-original-commit: 4480a3a337d0ff2915eb33e4bee3d4929886dc73
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Mathias Mathy (mama) <mama@odoo.com>