[FIX] helpdesk_timesheet: Fix validation error when creating a ticket
Steps to reproduce:
-------------------

- Install 'helpdesk_timesheet' module
- Go to Timesheets and click on `Start` to launch the timer
- Select `VIP Support` as Project
- Create a new ticket

Issue:
------

1. Validation error raised: `At least one analytic account must be set.`
2. When we select a helpdesk project that does not have any ticket
linked to it already, the field next to it is set to `Task` (task_id)
instead of `Ticket` (helpdesk_ticket_id).

Cause:
------

1. When creating the ticket, we don't provide the default project who
got the analytic account linked to it.
2. The field `helpdesk_ticket_id` is hidden because we consider that we
the project is not linked to any helpdesk team; this is because we
retrieve the projects based on tickets already created only.

Solution:
---------

1. Set the default project in the context when creating the ticket (like
for a task).
In the create method, if no team is provided and we have a default
project (from `list_value` or the `env.context`), we set the team based
on the project if the project is linked to a helpdesk team.
2. Fetch all projects that have `helpdesk_team` field set.

opw-4404097

closes odoo/enterprise#91279

X-original-commit: c2ce2c36e4c31efa34d8e254aaca5274cf391f91
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>