[IMP] documents_{*}: avoid file centralization empty settings block
When installing only documents, an empty section "Files centralization" is
present in the settings.

We solve the problem by making invisible by default when it is declared in
the documents App and make it visible in all module that add a setting in that
block.

Task-4294237

Part-of: odoo/enterprise#74571
Related: odoo/odoo#191459
Related: odoo/upgrade#6820
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>