[FIX] account_reports: report builder line out of sequence
When we have a report line that has a sequence lower than the sequence
of its parent, the line is not shown in report builder interface. This
prevents users to be able to fix the issue.

Also, the report will show a traceback.

This commit replaces the traceback with a UserError that states the
issue. It also shows and highlights the line that is out of sequence
so users can fix the issue.

closes odoo/enterprise#78010

X-original-commit: 6e5c8064854ee278ccf81532bc5ac3c5a4295002
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>