[FIX] knowledge: avoid opening new comments on click
This commit fixes an issue with the knowledge comments small UI.
When you create a comment in small UI and click on the overlay in the body of
the article, the newly created comments open its popover.

This behavior shouldn't happen in any circumstances except when in readonly
mode.

task-4463401

X-original-commit: 91cf63af2afcc8adc097d6f8bb9fac49914fe5eb
Part-of: odoo/enterprise#76898
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>