[REF] *: remove _get_rule_name method
The get rule name method was used as a workaround to get the translated
name of a rule. This method was a hack and a temporary solution. The
proper way to get the translated name of a rule is to use a real rule
with an XML ID. That way the user can translate the name of the rule
using the normal translation mechanism of Odoo.

task-3844686

Part-of: odoo/enterprise#59916
Related: odoo/upgrade#7046
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>