[FIX] test_sale_subscription: make uninstall possible
The uninstallation of this test module, which is a standard
step during upgrades, requires the removal of the test product
'product_fire_insurance'. This product is referenced by sale
orders and invoices and deferred invoices created dynamically
and have no xmlids, blocking the removal of the record.
This commits adds the xmlids to the dynamically created
orders and moves so they can be removed as well during the
module uninstallation.
The target sale orders are those created via 'copy' and when
doing upsell and renew for subscriptions.
The target moves are the regular invoices for the sale orders
and their deferred moves.

closes odoo/enterprise#77718

X-original-commit: 5d516f61c931c64c2bd171be79906b7ac10a127e
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>