[IMP] account_reports,l10n_*: Unambiguous cross-report aggregations
Before this commit, when using cross_report subformula it would search for a report that has an line with this code.
We had to ensure there was no duplicate codes and so it was difficult to maintain.

Now, when we use a cross_report subformula, we need to explicitly target the report from where the code is from.

task-4457667

closes odoo/enterprise#76845

Related: odoo/odoo#193106
Related: odoo/documentation#11747
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>