[IMP] documents_{*}: enable by default document centralization for some modules
We enable by default the document centralization for some modules.

Note that the post init is necessary as the module only set the folder for the
main company and other company can exist at the install of those module. A
simple way to test that is to install stock (which create
"My company (Chicago)") and documents_hr. We apply the same correction to
documents_fleet for which the document centralization was already enabled by
default.

Task-4294237

Part-of: odoo/enterprise#74571
Related: odoo/odoo#191459
Related: odoo/upgrade#6820
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>